backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
xgboost==2.1.1
pandas==2.2.2
numpy==1.26.4
scipy==1.13.1
requests==2.32.3
aiohttp==3.9.5
beautifulsoup4==4.12.3
lxml==5.2.2
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.7.4
scikit-learn==1.5.1
supabase==2.5.3
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
slowapi==0.1.9
sentry-sdk[fastapi]==2.8.0
apscheduler==3.10.4
resend==2.0.0
optuna==3.6.1
